refactor 빈 beandifinitions 변수명 변경 ;
